MAIN: Add dynamic cache system.

+++ IFR/main.py
@@ -1,7 +1,7 @@
 """User interface file for the Interference-Fringe-Removal program."""
 
 
-from definitions import FRINGE_CACHE_PATH, SIFG_CACHE_PATH, SSC_CACHE_PATH
+from definitions import FRINGE_CACHE_PATH, ROOT_DIR, SIFG_CACHE_PATH, SSC_CACHE_PATH
 from functools import partial
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
@@ -589,26 +589,49 @@ class Controller(object):
         np.savetxt(path[:-4] + "_REMOVED_FRINGES.dpt", fringe_locations, fmt="%4.7f", delimiter=",")
 
 
+def cache_setup(root_dir):
+    """Setup the cache file system.
+
+    This function initializes the programs file cache system of directories.
+    
+    Parameters
+    ----------
+    root_dir : str
+        string representing the programs root directory.
+    """
+
+    os.makedirs(root_dir + "/cache/fringe_spectrographs")
+    os.makedirs(root_dir + "/cache/SIFG_plot_data")
+    os.makedirs(root_dir + "/cache/SSC_plot_data")
+
 def program_exit():
     """Exit the UI program.
 
-    This function closes the UI after deleting all program cache files.
+    This function closes the UI after deleting all program cache files and the
+    cache directories.
     """
 
     app.exec()
 
     paths = [
+        FRINGE_CACHE_PATH + "/",
         SIFG_CACHE_PATH + "/",
-        SSC_CACHE_PATH + "/",
-        FRINGE_CACHE_PATH + "/"
+        SSC_CACHE_PATH + "/"
     ]
 
     for path in paths:
         files = os.listdir(path)
         for file in files:
-            os.remove(path + file)
+            if file[-4:] == ".npy":
+                os.remove(path + file)
+
+    os.rmdir(FRINGE_CACHE_PATH)
+    os.rmdir(SIFG_CACHE_PATH)
+    os.rmdir(SSC_CACHE_PATH)
+    os.rmdir(os.path.join(ROOT_DIR, "cache"))
 
 
+cache_setup(ROOT_DIR)
 app = QApplication([])
 app.setStyle("Windows")
 ui = UI()
